+++ test/e2e/app/run.go
@@ -7,7 +7,6 @@ import (
 	"github.com/omni-network/omni/lib/errors"
 	"github.com/omni-network/omni/lib/k1util"
 	"github.com/omni-network/omni/lib/log"
-	"github.com/omni-network/omni/lib/math"
 	"github.com/omni-network/omni/test/e2e/netman/pingpong"
 	"github.com/omni-network/omni/test/e2e/types"
 
@@ -133,7 +132,7 @@ func E2ETest(ctx context.Context, def Definition, cfg E2ETestConfig, depCfg Depl
 		return err
 	}
 
-	if err := WaitAllSubmissions(ctx, def.Netman.Portals(), math.Sum(msgBatches)); err != nil {
+	if err := WaitAllSubmissions(ctx, def.Netman.Portals(), Sum(msgBatches)); err != nil {
 		return err
 	}
